admin_backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
cryptography==41.0.7
httpx==0.26.0
bitcoinlib==0.6.14
loguru==0.7.2
python-multipart==0.0.6
